@@ -1,11 +1,9 @@
-import 'package:flutter/foundation.dart';
-
 class SpidIdentityProvider {
   final String id;
   final String name;
 
   const SpidIdentityProvider({
-    @required this.id,
-    @required this.name,
+    required this.id,
+    required this.name,
   });
 }
